arm64: WFI in cpu_relax()
cpu_relax() is used in the retry path of load-store exclusive loops to
allow another CPU to make progress before trying again.
Currently 'yield' is used, which is typically a NOP, but is a hint to
simluators that they should schedule another CPU.

Latency problems have been seen on non-LSE platforms where some CPUs
are unable to make progress when taking a lock because another CPU
always wins the race to hold the cache line in exclusive state.

Add WFI to the rety path to give other CPUs time to make progress.
This is only enabled after a second consecutive call to cpu_relax()
from the same call-site, and only if the eventstream is enabled.

CC: Will Deacon <will@kernel.org> # SOB for co-developed by?
Signed-off-by: James Morse <james.morse@arm.com>